Server/requirements.txt
Flask==3.0.0
Flask-SSE==0.2.0
Flask-SocketIO==5.3.6
celery==5.3.4
redis==5.0.1
python-socketio==5.10.0
eventlet==0.33.3
python-dotenv==1.0.0
Werkzeug==3.0.1

